+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 All Notable changes to `spatie-dom` will be documented in this file
 
-## 1.3.1
-- Fixes `hasTag` parameter order
+## 1.4.0 - 2017-05-15
+- Added `isInput`, `isTextInput`, `isCheckbox`, `isRadio` & `isTextarea` functions
 
-## 1.3.0
+## 1.3.1 - 2017-05-15
+- Fixed `hasTag` parameter order
+
+## 1.3.0 - 2017-05-08
 - Added `hasTag`
 
 ## 1.2.0